+++ docs/html_dev/search/all_10.js
@@ -1,8 +1,8 @@
 var searchData=
 [
   ['thousandsep',['ThousandSep',['../classxlib_1_1ThousandSep.html',1,'xlib']]],
-  ['time_5fstamp1',['time_stamp1',['../classcu__stinger_1_1Edge.html#ac90bd8b5f41c91401fc02b611321f04c',1,'cu_stinger::Edge']]],
-  ['time_5fstamp2',['time_stamp2',['../classcu__stinger_1_1Edge.html#a73ad2be660d612475143e6b05005702f',1,'cu_stinger::Edge']]],
+  ['time_5fstamp1',['time_stamp1',['../classcu__stinger_1_1Edge.html#ad40cc9724f3e2551e1f691bbfefb7baf',1,'cu_stinger::Edge']]],
+  ['time_5fstamp2',['time_stamp2',['../classcu__stinger_1_1Edge.html#ac8971eba5d8263e902a728bb635029bd',1,'cu_stinger::Edge']]],
   ['timer',['Timer',['../classtimer_1_1Timer.html',1,'timer']]],
   ['timer_2ecuh',['Timer.cuh',['../Timer_8cuh.html',1,'']]],
   ['timer_2ehpp',['Timer.hpp',['../Timer_8hpp.html',1,'']]],
@@ -13,5 +13,14 @@ var searchData=
   ['timerbase',['TimerBase',['../classtimer_1_1TimerBase.html',1,'timer']]],
   ['timerbase_3c_20cpu_2c_20chronoprecision_20_3e',['TimerBase&lt; CPU, ChronoPrecision &gt;',['../classtimer_1_1TimerBase.html',1,'timer']]],
   ['timerbase_3c_20device_2c_20chronoprecision_20_3e',['TimerBase&lt; DEVICE, ChronoPrecision &gt;',['../classtimer_1_1TimerBase.html',1,'timer']]],
-  ['timerbase_3c_20host_2c_20chronoprecision_20_3e',['TimerBase&lt; HOST, ChronoPrecision &gt;',['../classtimer_1_1TimerBase.html',1,'timer']]]
+  ['timerbase_3c_20host_2c_20chronoprecision_20_3e',['TimerBase&lt; HOST, ChronoPrecision &gt;',['../classtimer_1_1TimerBase.html',1,'timer']]],
+  ['tuple_5fcompare',['tuple_compare',['../structxlib_1_1tuple__compare.html',1,'xlib']]],
+  ['tuple_5fcompare_3c_20std_3a_3atuple_3c_20t1_2c_20targs1_2e_2e_2e_20_3e_2c_20std_3a_3atuple_3c_20t2_2c_20targs2_2e_2e_2e_20_3e_20_3e',['tuple_compare&lt; std::tuple&lt; T1, TArgs1... &gt;, std::tuple&lt; T2, TArgs2... &gt; &gt;',['../structxlib_1_1tuple__compare_3_01std_1_1tuple_3_01T1_00_01TArgs1_8_8_8_01_4_00_01std_1_1tuple_3_25abcc0c421bb76f455b2436c636537b.html',1,'xlib']]],
+  ['tuple_5fcompare_3c_20std_3a_3atuple_3c_3e_2c_20std_3a_3atuple_3c_3e_20_3e',['tuple_compare&lt; std::tuple&lt;&gt;, std::tuple&lt;&gt; &gt;',['../structxlib_1_1tuple__compare_3_01std_1_1tuple_3_4_00_01std_1_1tuple_3_4_01_4.html',1,'xlib']]],
+  ['tuple_5frm_5fpointers',['tuple_rm_pointers',['../structxlib_1_1tuple__rm__pointers.html',1,'xlib']]],
+  ['tuple_5frm_5fpointers_3c_20std_3a_3atuple_3c_20targs_2e_2e_2e_20_3e_20_3e',['tuple_rm_pointers&lt; std::tuple&lt; TArgs... &gt; &gt;',['../structxlib_1_1tuple__rm__pointers_3_01std_1_1tuple_3_01TArgs_8_8_8_01_4_01_4.html',1,'xlib']]],
+  ['tuple_5frm_5fpointers_5faux',['tuple_rm_pointers_aux',['../structxlib_1_1tuple__rm__pointers__aux.html',1,'xlib']]],
+  ['tuple_5frm_5fpointers_5faux_3c_20std_3a_3atuple_3c_20targs_2e_2e_2e_20_3e_2c_20std_3a_3atuple_3c_3e_20_3e',['tuple_rm_pointers_aux&lt; std::tuple&lt; TArgs... &gt;, std::tuple&lt;&gt; &gt;',['../structxlib_1_1tuple__rm__pointers__aux_3_01std_1_1tuple_3_01TArgs_8_8_8_01_4_00_01std_1_1tuple_3_4_01_4.html',1,'xlib']]],
+  ['tuple_5frm_5fpointers_5faux_3c_20std_3a_3atuple_3c_20targs1_2e_2e_2e_20_3e_2c_20std_3a_3atuple_3c_20t2_2c_20targs2_2e_2e_2e_20_3e_20_3e',['tuple_rm_pointers_aux&lt; std::tuple&lt; TArgs1... &gt;, std::tuple&lt; T2, TArgs2... &gt; &gt;',['../structxlib_1_1tuple__rm__pointers__aux_3_01std_1_1tuple_3_01TArgs1_8_8_8_01_4_00_01std_1_1tuple_145804f0bb247f90dd3bcb76b0762148.html',1,'xlib']]],
+  ['tuple_5frm_5fpointers_5faux_3c_20std_3a_3atuple_3c_20targs1_2e_2e_2e_2c_20std_3a_3aremove_5fpointer_3c_20t2_20_3e_3a_3atype_20_3e_2c_20std_3a_3atuple_3c_20targs2_2e_2e_2e_20_3e_20_3e',['tuple_rm_pointers_aux&lt; std::tuple&lt; TArgs1..., std::remove_pointer&lt; T2 &gt;::type &gt;, std::tuple&lt; TArgs2... &gt; &gt;',['../structxlib_1_1tuple__rm__pointers__aux.html',1,'xlib']]]
 ];