+++ docs/html_dev/search/classes_0.js
@@ -1,6 +1,7 @@
 var searchData=
 [
-  ['batchproperty',['BatchProperty',['../structBatchProperty.html',1,'']]],
+  ['batchproperty',['BatchProperty',['../classcu__stinger_1_1BatchProperty.html',1,'cu_stinger::BatchProperty'],['../structBatchProperty.html',1,'BatchProperty']]],
+  ['batchupdate',['BatchUpdate',['../classcu__stinger_1_1BatchUpdate.html',1,'cu_stinger']]],
   ['binomialcoeff',['BinomialCoeff',['../structxlib_1_1BinomialCoeff.html',1,'xlib']]],
   ['bittree',['BitTree',['../classcu__stinger_1_1BitTree.html',1,'cu_stinger']]]
 ];