+++ docs/html_dev/search/functions_0.js
@@ -1,6 +1,8 @@
 var searchData=
 [
   ['base_5faddress',['base_address',['../classcu__stinger_1_1BitTree.html#a93583462a165a9f183b65a52722bd0b7',1,'cu_stinger::BitTree']]],
+  ['batchproperty',['BatchProperty',['../classcu__stinger_1_1BatchProperty.html#a973819b8515b2f5b8beea55e47b54d35',1,'cu_stinger::BatchProperty']]],
+  ['batchupdate',['BatchUpdate',['../classcu__stinger_1_1BatchUpdate.html#aaf793aafaa1151c32fa6d6f447b2d09c',1,'cu_stinger::BatchUpdate']]],
   ['belong_5fto',['belong_to',['../classcu__stinger_1_1BitTree.html#ab4c93353aa03bd11b10f98b3ee7cd344',1,'cu_stinger::BitTree']]],
   ['bittree',['BitTree',['../classcu__stinger_1_1BitTree.html#a035e893a4c89565378fcb6c39d2974ee',1,'cu_stinger::BitTree::BitTree() noexcept'],['../classcu__stinger_1_1BitTree.html#a719c823a9b9debd90a2b12cea2612ca7',1,'cu_stinger::BitTree::BitTree(BitTree &amp;&amp;obj) noexcept']]]
 ];