+++ docs/html_dev/search/functions_10.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['weight',['weight',['../classcu__stinger_1_1Edge.html#aab600b207308f8988743ee1110758d40',1,'cu_stinger::Edge']]]
+  ['weight',['weight',['../classcu__stinger_1_1Edge.html#afe9308bb9e5a877ea428f14c222df827',1,'cu_stinger::Edge']]]
 ];