+++ docs/html_dev/search/functions_6.js
@@ -1,8 +1,7 @@
 var searchData=
 [
-  ['initialize',['initialize',['../classcu__stinger_1_1cuStinger.html#a9c755a111214652f24904927205d8fa4',1,'cu_stinger::cuStinger']]],
   ['insert',['insert',['../classcu__stinger_1_1BitTree.html#ab826805c6098d205325d9064bbed8c66',1,'cu_stinger::BitTree::insert()'],['../classcu__stinger_1_1MemoryManagement.html#a0b783ab596aecb4fb0f815ac3686c89a',1,'cu_stinger::MemoryManagement::insert()']]],
-  ['insertedgedata',['insertEdgeData',['../classcu__stinger_1_1cuStinger.html#aff24804befe4ebe910d1bbd76a617efc',1,'cu_stinger::cuStinger']]],
-  ['insertvertexdata',['insertVertexData',['../classcu__stinger_1_1cuStinger.html#a85b9c28a862a05443e44babe8d6eca28',1,'cu_stinger::cuStinger']]],
+  ['insertedgedata',['insertEdgeData',['../classcu__stinger_1_1cuStingerInit.html#a0c7e20e31f020ad78f8757771f0160d3',1,'cu_stinger::cuStingerInit::insertEdgeData()'],['../classcu__stinger_1_1BatchUpdate.html#ad5f39a6b11d9d92308eaca72c6c302e9',1,'cu_stinger::BatchUpdate::insertEdgeData()']]],
+  ['insertvertexdata',['insertVertexData',['../classcu__stinger_1_1cuStingerInit.html#ac77b6434bfa10d12acdde70abeb27711',1,'cu_stinger::cuStingerInit']]],
   ['is_5ffull',['is_full',['../classcu__stinger_1_1BitTree.html#a169f276f03124022c8047de24a48dda6',1,'cu_stinger::BitTree']]]
 ];