+++ docs/html_dev/search/functions_e.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['time_5fstamp1',['time_stamp1',['../classcu__stinger_1_1Edge.html#ac90bd8b5f41c91401fc02b611321f04c',1,'cu_stinger::Edge']]],
-  ['time_5fstamp2',['time_stamp2',['../classcu__stinger_1_1Edge.html#a73ad2be660d612475143e6b05005702f',1,'cu_stinger::Edge']]]
+  ['time_5fstamp1',['time_stamp1',['../classcu__stinger_1_1Edge.html#ad40cc9724f3e2551e1f691bbfefb7baf',1,'cu_stinger::Edge']]],
+  ['time_5fstamp2',['time_stamp2',['../classcu__stinger_1_1Edge.html#ac8971eba5d8263e902a728bb635029bd',1,'cu_stinger::Edge']]]
 ];